% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2dropSmart-package.R
\docType{package}
\name{r2dropSmart-package}
\alias{r2dropSmart}
\alias{r2dropSmart-package}
\title{r2dropSmart}
\description{
Sync entire folder with Dropbox.
Use of cache to update only new files or changed files.
}
