% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{sync}
\alias{sync}
\title{Syncronize full folder to dropbox path}
\usage{
sync(path, remote, token = NULL, pattern = NULL, blackList = NULL,
  share = FALSE, dry = FALSE, force = FALSE, ...)
}
\arguments{
\item{path}{Local path in the computer}

\item{remote}{Remote path in dropbox}

\item{token}{Security token. \code{\link[rdrop2:drop_auth]{rdrop2::drop_auth()}}}

\item{pattern}{Expression pattern to keep only those files.
Pass into \code{\link[base:list.files]{base::list.files()}}.}

\item{blackList}{Character vector that will be used to remove
files that match those expression.}

\item{share}{Whether create a sharing link or not.}

\item{dry}{Not perform action but list the files to be upload.}

\item{force}{Force to upload files to drobox.}

\item{...}{Options to pass to \code{\link[rdrop2:drop_share]{rdrop2::drop_share()}}}
}
\description{
Copy an entire folder to dropbox. It allows filtering by
pattern or/and blacklist.
It optionally creates the sharing link.
It upload only changed files, using \code{\link[tools:md5sum]{tools::md5sum()}} to
compare versions.
}
\examples{

local <- "."
remote <- "path/to/drobox"
library(rdrop2)
# token <- readRDS("~/.droptoken.rds")
# sync(loca, remote, pattern = ".R", dry = TRUE)
}
